+++ Controller/TrendsController.php
@@ -35,11 +35,9 @@ class TrendsController extends Controller
 
     /**
      * @Route("/timings", name="dtc_queue_timings")
-     *
      */
-    public function getTimingsAction()
+    public function getTimingsAction(Request $request)
     {
-        $request = $this->get('request_stack')->getMasterRequest();
         $begin = $request->query->get('begin');
         $end = $request->query->get('end');
         $type = $request->query->get('type', 'HOUR');
@@ -57,7 +55,7 @@ class TrendsController extends Controller
 
     /**
      * @param \DateTime|null $beginDate
-     * @param \DateTime $endDate
+     * @param \DateTime      $endDate
      */
     protected function calculateTimings($type, $beginDate, $endDate)
     {
@@ -83,7 +81,7 @@ class TrendsController extends Controller
         }
 
         $format = $this->getDateFormat($type);
-        usort($timingsDates, function($date1str, $date2str) use ($format) {
+        usort($timingsDates, function ($date1str, $date2str) use ($format) {
             $date1 = \DateTime::createFromFormat($format, $date1str);
             $date2 = \DateTime::createFromFormat($format, $date2str);
             if (!$date2) {
@@ -107,7 +105,7 @@ class TrendsController extends Controller
     /**
      * Timings offset by timezone if necessary.
      *
-     * @param array $timingsDates
+     * @param array  $timingsDates
      * @param string $format
      *
      * @return array
@@ -120,6 +118,9 @@ class TrendsController extends Controller
             $date = \DateTime::createFromFormat($format, $dateStr);
             if (0 !== $timezoneOffset) {
                 $date->setTimestamp($date->getTimestamp() + ($timezoneOffset * 3600));
+            }
+            if (false === $date) {
+                throw new \InvalidArgumentException("'$date' is not in the right format: ".DATE_RFC3339);
             }
             $timingsDatesAdjusted[] = $date->format(DATE_RFC3339);
         }
@@ -157,11 +158,11 @@ class TrendsController extends Controller
         }
 
         // Run a map reduce function get worker and status break down
-        $mapFunc = "function() {
+        $mapFunc = 'function() {
             var dateStr = this.finishedAt.toISOString();
-            dateStr = dateStr.replace(/{$regexInfo['regex']}/,'{$regexInfo['replacement']}');
-            var dateBegin = new Date('{$begin->format('c')}');
-            var dateEnd = new Date('{$end->format('c')}');
+            dateStr = dateStr.replace(/'.$regexInfo['replace']['regex']."/,'".$regexInfo['replace']['replacement']."');
+            var dateBegin = new Date('".$begin->format('c')."');
+            var dateEnd = new Date('".$end->format('c')."');
             if (this.finishedAt >= dateBegin && this.finishedAt <= dateEnd) {
                 var result = {};
                 result[dateStr] = 1;
@@ -273,10 +274,43 @@ class TrendsController extends Controller
 
         $resultHash = [];
         foreach ($result as $row) {
-            $resultHash[$row['status']][$row['thedate']] = intval($row['thecount']);
+            $date = $this->formatOrmDateTime($type, $row['thedate']);
+            $resultHash[$row['status']][$date] = intval($row['thecount']);
         }
 
         return $resultHash;
+    }
+
+    protected function strPadLeft($str)
+    {
+        return str_pad($str, 2, '0', STR_PAD_LEFT);
+    }
+
+    protected function formatOrmDateTime($type, $str)
+    {
+        switch ($type) {
+            case 'MONTH':
+                $parts = explode('-', $str);
+
+                return $parts[0].'-'.$this->strPadLeft($parts[1]);
+            case 'DAY':
+                $parts = explode('-', $str);
+
+                return $parts[0].'-'.$this->strPadLeft($parts[1]).'-'.$this->strPadLeft($parts[2]);
+            case 'HOUR':
+                $parts = explode(' ', $str);
+                $dateParts = explode('-', $parts[0]);
+
+                return $dateParts[0].'-'.$this->strPadLeft($dateParts[1]).'-'.$this->strPadLeft($dateParts[2]).' '.$this->strPadLeft($parts[1]);
+            case 'MINUTE':
+                $parts = explode(' ', $str);
+                $dateParts = explode('-', $parts[0]);
+                $timeParts = explode(':', $parts[1]);
+
+                return $dateParts[0].'-'.$this->strPadLeft($dateParts[1]).'-'.$this->strPadLeft($dateParts[2]).' '.$this->strPadLeft($timeParts[0]).':'.$this->strPadLeft($timeParts[1]);
+        }
+
+        return $str;
     }
 
     protected function validateJobTimingManager()