+++ ODM/JobManager.php
@@ -15,7 +15,7 @@ class JobManager extends BaseJobManager
     const REDUCE_FUNCTION = 'function(k, vals) {
             var result = {};
             for (var index in vals) {
-                var val =  vals[index];
+                var val = vals[index];
                 for (var i in val) {
                     if (result.hasOwnProperty(i)) {
                         result[i] += val[i];
@@ -383,8 +383,8 @@ class JobManager extends BaseJobManager
     }
 
     /**
-     * @param string $workerName
-     * @param string $methodName
+     * @param string   $workerName
+     * @param string   $methodName
      * @param \Closure $progressCallback
      */
     public function archiveAllJobs($workerName = null, $methodName = null, $progressCallback)